# Improve performance of SplitLocale