--- a/.github/workflows/test.yaml
+++ b/.github/workflows/test.yaml
@@ -22,11 +22,11 @@ jobs:
       matrix:
         os: [ubuntu-latest, macos-latest, windows-latest]
         go:
-          - ~1.14
-          - ~1.15
-          - ~1.16
           - ~1.17
           - ~1.18
+          - ~1.19
+          - ~1.20
+          - ~1.21
     runs-on: ${{ matrix.os }}
     steps:
       - name: Checkout the code
@@ -36,6 +36,7 @@ jobs:
         with:
           go-version: ${{ matrix.go }}
       - name: Cache the Go modules
+        id: cache
         uses: actions/cache@v3
         with:
           path: ~/go/pkg/mod
@@ -78,7 +79,7 @@ jobs:
         os: [ubuntu-latest, macos-latest, windows-latest]
     runs-on: ${{ matrix.os }}
     env:
-      GOLANGCI_LINT_VERSION: 1.46.2
+      GOLANGCI_LINT_VERSION: 1.55.2
     steps:
       - name: Checkout the code
         uses: actions/checkout@v3