--- a/.golangci.yml
+++ b/.golangci.yml
@@ -4,16 +4,13 @@ linters:
   disable-all: true
   enable:
     # Default
-    - deadcode
     - errcheck
     - gosimple
     - govet
     - ineffassign
     - staticcheck
-    - structcheck
     - typecheck
     - unused
-    - varcheck
     # Manually enabled
     - dupl
     - gocyclo