--- a/examples/getlocale-gui/android.go
+++ b/examples/getlocale-gui/android.go
@@ -4,11 +4,16 @@
 package main
 
 import (
-	"github.com/fyne-io/mobile/app"
+	"fyne.io/fyne/v2/driver"
 
 	"github.com/jeandeaual/go-locale"
 )
 
 func init() {
-	locale.SetRunOnJVM(app.RunOnJVM)
+	locale.SetRunOnJVM(func(fn func(vm, env, ctx uintptr) error) {
+		driver.RunNative(func(ctx interface{}) error {
+			and := env.(driver.AndroidContext)
+			fn(and.VM, and.Env, and.Ctx)
+		})
+	})
 }