--- a/go.mod
+++ b/go.mod
@@ -3,10 +3,9 @@ module github.com/jeandeaual/go-locale
 go 1.14
 
 require (
-	fyne.io/fyne v1.4.2
-	github.com/fyne-io/mobile v0.1.2
+	fyne.io/fyne/v2 v2.4.3
 	github.com/nicksnyder/go-i18n/v2 v2.1.1
-	github.com/stretchr/testify v1.6.1
-	golang.org/x/sys v0.0.0-20211013075003-97ac67df715c
-	golang.org/x/text v0.3.4
+	github.com/stretchr/testify v1.8.4
+	golang.org/x/sys v0.13.0
+	golang.org/x/text v0.13.0
 )