--- a/go.mod
+++ b/go.mod
@@ -4,8 +4,8 @@ go 1.14
 
 require (
 	fyne.io/fyne/v2 v2.4.3
-	github.com/nicksnyder/go-i18n/v2 v2.1.1
+	github.com/nicksnyder/go-i18n/v2 v2.4.0
 	github.com/stretchr/testify v1.8.4
-	golang.org/x/sys v0.13.0
-	golang.org/x/text v0.13.0
+	golang.org/x/sys v0.16.0
+	golang.org/x/text v0.14.0
 )