--- a/locale_darwin.go
+++ b/locale_darwin.go
@@ -49,11 +49,7 @@ func GetLocale() (string, error) {
 
 // appleLanguagesRegex is used to parse the output of "defaults read -g AppleLanguages"
 // e.g.:
-// (
-//     en,
-//     "fr-FR",
-//     "ja-JP"
-// )
+// (en, "fr-FR", "ja-JP")
 var appleLanguagesRegex = regexp.MustCompile(`([a-z]{2}(?:-[A-Z]{2})?)`)
 
 // GetLocales retrieves the IETF BCP 47 language tags set on the system.