This improves performance and makes the code more readable.

Benchstat:
```
goos: linux
goarch: amd64
pkg: github.com/fyne-io/go-locale
cpu: Intel(R) Core(TM) i7-7700 CPU @ 3.60GHz
              │   old.txt    │              new.txt               │
              │    sec/op    │   sec/op     vs base               │
SplitLocale-8   119.75n ± 1%   20.55n ± 1%  -82.84% (p=0.002 n=6)

              │  old.txt   │              new.txt              │
              │    B/op    │   B/op     vs base                │
SplitLocale-8   64.00 ± 0%   0.00 ± 0%  -100.00% (p=0.002 n=6)

              │  old.txt   │              new.txt               │
              │ allocs/op  │ allocs/op   vs base                │
SplitLocale-8   2.000 ± 0%   0.000 ± 0%  -100.00% (p=0.002 n=6)
```