@Jacalz 2024-02-19
Closing in favour of https://github.com/jeandeaual/go-locale/pull/23